+++ nautilus_trader/model/instruments/__init__.py
@@ -27,6 +27,7 @@ from nautilus_trader.model.instruments.equity import Equity
 from nautilus_trader.model.instruments.futures_contract import FuturesContract
 from nautilus_trader.model.instruments.futures_spread import FuturesSpread
 from nautilus_trader.model.instruments.options_contract import OptionsContract
+from nautilus_trader.model.instruments.options_spread import OptionsSpread
 from nautilus_trader.model.instruments.synthetic import SyntheticInstrument
 
 
@@ -40,6 +41,7 @@ __all__ = [
     "FuturesContract",
     "FuturesSpread",
     "OptionsContract",
+    "OptionsSpread",
     "SyntheticInstrument",
     "instruments_from_pyo3",
 ]